Fixed Digital mode of AGC2 implementation finished.
This CL adds the GainCurveApplier (GCA). It owns a
FixedDigitalLevelEstimator (LE) and an InterpolatedGainCurve
(IGC). The GCA uses the LE to compute the input signal level, looks up
a gain from IGC and applies it on the signal.

The other IGC and LE submodules were added in previous CLs [1] and
[2].

This CL also turns on AGC2 in the APM fuzzer.

[1] https://webrtc-review.googlesource.com/c/src/+/51920
[2] https://webrtc-review.googlesource.com/c/src/+/52381


Bug: webrtc:7949

+++ modules/audio_processing/agc2/agc2_testing_common.h
@@ -23,9 +23,9 @@ namespace test {
 constexpr float kDecayMs = 500.f;
 
 // Limiter parameters.
-constexpr double kLimiterMaxInputLevel = 1.0;
-constexpr double kLimiterKneeSmoothness = 1.0;
-constexpr double kLimiterCompressionRatio = 5.0;
+constexpr float kLimiterMaxInputLevel = 1.f;
+constexpr float kLimiterKneeSmoothness = 1.f;
+constexpr float kLimiterCompressionRatio = 5.f;
 
 std::vector<double> LinSpace(const double l, const double r, size_t num_points);
 }  // namespace test
